Phase 0.9: Linting warnings fix
